Compute the Radial Electric Field (Er).
Adds computation of the radial electric field and stores it in CoreProfiles. This will be used to compute ExB shear for transport models in a follow-up PR.

For this computation, we need to know the toroidal and poloidal velocity of the plasma. We add a profile condition for specifying the toroidal velocity, which needs to be prescribed, otherwise it will be set to 0. The poloidal velocity can be computed by the neoclassical model, to be implemented in a follow-up PR.

PiperOrigin-RevId: 826167942