Make pressures cached properties of core_profiles.
This will help avoid duplicated computation/definitions of the pressure throughout the codebase. This current change doesn't change anything in terms of physics computation and outputs. In a follow-up CL, we can update the sauter and angioni-sauter models to use the pressure from core_profiles. This will change the physics a bit, since these models currently ignore the ion pressure coming from impurities. We decouple these change from the internal core profiles change in order to better track the changes in the outputs.

PiperOrigin-RevId: 832313095