Make ToraxSimState immutable.
This change makes `ToraxSimState` and `SolverNumericOutputs` frozen dataclasses. This makes it more consistent with how we deal with dataclasses across TORAX. The downside is that modifying the state is now a bit more cumbersome. But this should not be needed very often. Apparently it is currently only done once in the output code.

PiperOrigin-RevId: 834906844