Impose consistency between edge and core impurities.
core and edge impurities are no longer independent parameters. They are coupled and consistent with each other, through the enrichment_factor.

Extensive validation of new logic.

Sets the source of truth for impurity densities when using edge models. When SoT is EDGE, then the core impurity boundary condition is set by the edge fixed impurity density in the edge config, divided by the enrichment factor. When SoT is CORE, then the user can set directly the core boundary condition in PlasmaComposition and the edge value is multiplied by the enrichment factor. Both options have their use-cases.

Note that regardless, the 1D profile shape in the core is set in PlasmaComposition. Also note that regardless, when extended-lengyel is used in INVERSE mode, then the seeded impurity subset has its SoT in the edge.

PiperOrigin-RevId: 837560095