Implement rotation rule for QLKNN transport models.
This follows the description of the rotation rule from the original QLKNN paper (van de Plassche K.L. 2020). The application of the rotation rule is gated by a flag that is false by default.

Validation of this rule will be required for both QLKNN10D and QLKNN_7_11 before we can enable it by default. It is possible that the rule cannot be directly applied to QLKNN_7_11 and requires further work.

PiperOrigin-RevId: 842207600