Improve geometric parameter resolution in ExtendedLengyelModel.
Needed because MEQ postprocessing does not work for all geometries and fallbacks are needed for robustness.

The `_resolve_param` helper now checks if the value provided by the Geometry object is valid (non-zero and not NaN). If the Geometry value is invalid, it falls back to the value provided in the ExtendedLengyelConfig. If no fallback is available in the Config, an error is raised.

PiperOrigin-RevId: 842254161